compile: Extract TypeInferrer as a separate class
Also collect referenced variables as a separate pass so that
TypeInferrer now only has a single responsibility and a more clear
contract.
